220812-GAN-for-Breast-Histopathology-Images
This is an update to prof Abdulhamit of my resent progress on learning from breast histopathology images. By refering to examples on Kaggle of GAN, a GAN for histopathology image dataset BHI is created. This will be used to later train model and see effect of GAN as image enhancement